Do not redownload

+++ install_into_usb_connected_wemos_d1.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
-curl -fsSL https://raw.githubusercontent.com/platformio/platformio-core-installer/master/get-platformio.py -o get-platformio.py
+if [ ! -f get-platformio.py ]; then
+	curl -fsSL https://raw.githubusercontent.com/platformio/platformio-core-installer/master/get-platformio.py -o get-platformio.py
+fi
+
 python3 get-platformio.py
 
 cd fingerprint-mqtt && $HOME/.platformio/penv/bin/platformio run -t upload
